initial builder refactoring
