+++ tests/Builder/BuildPayloadValidationTest.php
@@ -3,7 +3,7 @@
 
 declare(strict_types=1);
 
-namespace FppTest\Helpers;
+namespace FppTest\Builder;
 
 use Fpp\Argument;
 use Fpp\Constructor;
