+++ tests/Builder/BuildScalarConstructorTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace FppTest\Helpers;
+namespace FppTest\Builder;
 
 use Fpp\Constructor;
 use Fpp\Definition;
