+++ tests/Builder/BuildStaticConstructorBodyConvertingToPayloadTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace FppTest\Helpers;
+namespace FppTest\Builder;
 
 use Fpp\Argument;
 use Fpp\Constructor;
