+++ tests/ParseTest.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace FppTest;
 
+use function Fpp\defaultDerivingMap;
 use Fpp\Deriving;
 use Fpp\ParseError;
 use org\bovigo\vfs\vfsStream;
@@ -21,7 +22,7 @@ class ParseTest extends TestCase
     /**
      * @var array
      */
-    private $derivingsMap = [];
+    private $derivingMap = [];
 
     protected function setUp()
     {
@@ -29,21 +30,7 @@ class ParseTest extends TestCase
         vfsStream::newFile('not_readable.fpp')->withContent('')->at($this->root);
         $this->root->getChild('not_readable.fpp')->chmod('0000');
 
-        $this->derivingsMap = [
-            'AggregateChanged' => new Deriving\AggregateChanged(),
-            'Command' => new Deriving\Command(),
-            'DomainEvent' => new Deriving\DomainEvent(),
-            'Enum' => new Deriving\Enum(),
-            'Equals' => new Deriving\Equals(),
-            'FromArray' => new Deriving\FromArray(),
-            'FromScalar' => new Deriving\FromScalar(),
-            'FromString' => new Deriving\FromString(),
-            'Query' => new Deriving\Query(),
-            'ToArray' => new Deriving\ToArray(),
-            'ToScalar' => new Deriving\ToScalar(),
-            'ToString' => new Deriving\ToString(),
-            'Uuid' => new Deriving\Uuid(),
-        ];
+        $this->derivingMap = defaultDerivingMap();
     }
 
     protected function createDefaultFile(string $contents): string
@@ -60,7 +47,7 @@ class ParseTest extends TestCase
     {
         $this->expectException(\RuntimeException::class);
 
-        parse($this->root->url() . '/invalid', $this->derivingsMap);
+        parse($this->root->url() . '/invalid', $this->derivingMap);
     }
 
     /**
@@ -70,7 +57,7 @@ class ParseTest extends TestCase
     {
         $this->expectException(\RuntimeException::class);
 
-        parse($this->root->url() . '/not_readable.fpp', $this->derivingsMap);
+        parse($this->root->url() . '/not_readable.fpp', $this->derivingMap);
     }
 
     /**
@@ -79,7 +66,7 @@ class ParseTest extends TestCase
     public function it_parses_empty_file(): void
     {
         $contents = '';
-        $collection = parse($this->createDefaultFile($contents), $this->derivingsMap);
+        $collection = parse($this->createDefaultFile($contents), $this->derivingMap);
 
         $this->assertCount(0, $collection->definitions());
     }
@@ -90,7 +77,7 @@ class ParseTest extends TestCase
     public function it_parses_file_with_only_whitespace(): void
     {
         $contents = '     ';
-        $collection = parse($this->createDefaultFile($contents), $this->derivingsMap);
+        $collection = parse($this->createDefaultFile($contents), $this->derivingMap);
 
         $this->assertCount(0, $collection->definitions());
     }
@@ -103,7 +90,7 @@ class ParseTest extends TestCase
         $this->expectException(ParseError::class);
 
         $contents = 'namespace Something Invalid';
-        parse($this->createDefaultFile($contents), $this->derivingsMap);
+        parse($this->createDefaultFile($contents), $this->derivingMap);
     }
 
     /**
@@ -114,7 +101,7 @@ class ParseTest extends TestCase
         $this->expectException(ParseError::class);
 
         $contents = 'namespace Something{ namespace Foo; }';
-        parse($this->createDefaultFile($contents), $this->derivingsMap);
+        parse($this->createDefaultFile($contents), $this->derivingMap);
     }
 
     /**
@@ -125,7 +112,7 @@ class ParseTest extends TestCase
         $this->expectException(ParseError::class);
 
         $contents = 'namespace Something ';
-        parse($this->createDefaultFile($contents), $this->derivingsMap);
+        parse($this->createDefaultFile($contents), $this->derivingMap);
     }
 
     /**
@@ -137,7 +124,7 @@ class ParseTest extends TestCase
         $contents = <<<CODE
 namespace \;
 CODE;
-        parse($this->createDefaultFile($contents), $this->derivingsMap);
+        parse($this->createDefaultFile($contents), $this->derivingMap);
     }
 
     /**
@@ -149,7 +136,7 @@ CODE;
 namespace Something;
 data Name = String;
 CODE;
-        $collection = parse($this->createDefaultFile($contents), $this->derivingsMap);
+        $collection = parse($this->createDefaultFile($contents), $this->derivingMap);
 
         $this->assertTrue($collection->hasDefinition('Something', 'Name'));
 
@@ -169,7 +156,7 @@ CODE;
         $contents = <<<CODE
 data Name = String;
 CODE;
-        $collection = parse($this->createDefaultFile($contents), $this->derivingsMap);
+        $collection = parse($this->createDefaultFile($contents), $this->derivingMap);
 
         $this->assertTrue($collection->hasDefinition('', 'Name'));
         $definition = $collection->definition('', 'Name');
@@ -192,7 +179,7 @@ namespace Bar {
     data Name = String;
 }
 CODE;
-        $collection = parse($this->createDefaultFile($contents), $this->derivingsMap);
+        $collection = parse($this->createDefaultFile($contents), $this->derivingMap);
 
         $this->assertTrue($collection->hasDefinition('Foo', 'Name'));
         $this->assertTrue($collection->hasDefinition('Bar', 'Name'));
@@ -211,7 +198,7 @@ namespace Foo {
 }
 }
 CODE;
-        parse($this->createDefaultFile($contents), $this->derivingsMap);
+        parse($this->createDefaultFile($contents), $this->derivingMap);
     }
 
     /**
@@ -225,7 +212,7 @@ CODE;
 namespace Something\Here;
 data Name = String
 CODE;
-        parse($this->createDefaultFile($contents), $this->derivingsMap);
+        parse($this->createDefaultFile($contents), $this->derivingMap);
     }
 
     /**
@@ -239,7 +226,7 @@ CODE;
 namespace Something;
 data Name =;
 CODE;
-        parse($this->createDefaultFile($contents), $this->derivingsMap);
+        parse($this->createDefaultFile($contents), $this->derivingMap);
     }
 
     /**
@@ -253,7 +240,7 @@ CODE;
 namespace Something;
 data;
 CODE;
-        parse($this->createDefaultFile($contents), $this->derivingsMap);
+        parse($this->createDefaultFile($contents), $this->derivingMap);
     }
 
     /**
@@ -267,7 +254,7 @@ CODE;
 namespace Something;
 data = ;
 CODE;
-        parse($this->createDefaultFile($contents), $this->derivingsMap);
+        parse($this->createDefaultFile($contents), $this->derivingMap);
     }
 
     /**
@@ -281,7 +268,7 @@ CODE;
 namespace Something;
 data Person Person;
 CODE;
-        parse($this->createDefaultFile($contents), $this->derivingsMap);
+        parse($this->createDefaultFile($contents), $this->derivingMap);
     }
 
     /**
@@ -294,7 +281,7 @@ CODE;
         $contents = <<<CODE
 =
 CODE;
-        parse($this->createDefaultFile($contents), $this->derivingsMap);
+        parse($this->createDefaultFile($contents), $this->derivingMap);
     }
 
     /**
@@ -307,7 +294,7 @@ CODE;
         $contents = <<<CODE
 function
 CODE;
-        parse($this->createDefaultFile($contents), $this->derivingsMap);
+        parse($this->createDefaultFile($contents), $this->derivingMap);
     }
 
     /**
@@ -319,7 +306,7 @@ CODE;
 namespace Something;
 data Name = Name | FirstName;
 CODE;
-        $collection = parse($this->createDefaultFile($contents), $this->derivingsMap);
+        $collection = parse($this->createDefaultFile($contents), $this->derivingMap);
 
         $this->assertTrue($collection->hasDefinition('Something', 'Name'));
 
@@ -345,7 +332,7 @@ namespace Something;
 data Name = Name | FirstName;
 data Age = Int;
 CODE;
-        $collection = parse($this->createDefaultFile($contents), $this->derivingsMap);
+        $collection = parse($this->createDefaultFile($contents), $this->derivingMap);
 
         $this->assertTrue($collection->hasDefinition('Something', 'Name'));
         $this->assertTrue($collection->hasDefinition('Something', 'Age'));
@@ -376,7 +363,7 @@ CODE;
 namespace Something;
 invalid Name = String;
 CODE;
-        parse($this->createDefaultFile($contents), $this->derivingsMap);
+        parse($this->createDefaultFile($contents), $this->derivingMap);
     }
 
     /**
@@ -389,7 +376,7 @@ CODE;
 namespace Something;
 data Name = string;
 CODE;
-        parse($this->createDefaultFile($contents), $this->derivingsMap);
+        parse($this->createDefaultFile($contents), $this->derivingMap);
     }
 
     /**
@@ -402,7 +389,7 @@ namespace Something;
 data Person = Person { string \$name, ?int \$age } ;
 CODE;
 
-        $collection = parse($this->createDefaultFile($contents), $this->derivingsMap);
+        $collection = parse($this->createDefaultFile($contents), $this->derivingMap);
         $definition = $collection->definition('Something', 'Person');
         $constructor = $definition->constructors()[0];
         $this->assertCount(2, $constructor->arguments());
@@ -428,7 +415,7 @@ namespace Something;
 data Person = Person { string \$name, ?int \$age } | Chef { string \$name };
 CODE;
 
-        $collection = parse($this->createDefaultFile($contents), $this->derivingsMap);
+        $collection = parse($this->createDefaultFile($contents), $this->derivingMap);
         $definition = $collection->definition('Something', 'Person');
 
         $constructor1 = $definition->constructors()[0];
@@ -463,7 +450,7 @@ namespace Something;
 data Person = Person { string \$name, ?int \$age } | Chef;
 CODE;
 
-        $collection = parse($this->createDefaultFile($contents), $this->derivingsMap);
+        $collection = parse($this->createDefaultFile($contents), $this->derivingMap);
         $definition = $collection->definition('Something', 'Person');
 
         $constructor1 = $definition->constructors()[0];
@@ -493,7 +480,7 @@ namespace Something;
 data Person = Person { Data\Name \$name }; 
 CODE;
 
-        $collection = parse($this->createDefaultFile($contents), $this->derivingsMap);
+        $collection = parse($this->createDefaultFile($contents), $this->derivingMap);
         $definition = $collection->definition('Something', 'Person');
 
         $constructor = $definition->constructors()[0];
@@ -515,7 +502,7 @@ namespace Something;
 data Person = Person { \Data\Name \$name }; 
 CODE;
 
-        $collection = parse($this->createDefaultFile($contents), $this->derivingsMap);
+        $collection = parse($this->createDefaultFile($contents), $this->derivingMap);
         $definition = $collection->definition('Something', 'Person');
 
         $constructor = $definition->constructors()[0];
@@ -538,7 +525,7 @@ namespace My {
 }
 CODE;
 
-        $collection = parse($this->createDefaultFile($contents), $this->derivingsMap);
+        $collection = parse($this->createDefaultFile($contents), $this->derivingMap);
         $definition = $collection->definition('My', 'Color');
 
         $this->assertCount(2, $definition->constructors());
@@ -558,7 +545,7 @@ namespace My {
 }
 CODE;
 
-        $collection = parse($this->createDefaultFile($contents), $this->derivingsMap);
+        $collection = parse($this->createDefaultFile($contents), $this->derivingMap);
         $definition = $collection->definition('My', 'Person');
 
         $this->assertCount(1, $definition->constructors());
@@ -584,7 +571,7 @@ namespace Something;
 data Person = Person { string name, ?int \$age } ;
 CODE;
 
-        parse($this->createDefaultFile($contents), $this->derivingsMap);
+        parse($this->createDefaultFile($contents), $this->derivingMap);
     }
 
     /**
@@ -597,7 +584,7 @@ namespace Something;
 data Person = Person { string \$name, ?int \$age } deriving (ToArray, FromArray);
 CODE;
 
-        $collection = parse($this->createDefaultFile($contents), $this->derivingsMap);
+        $collection = parse($this->createDefaultFile($contents), $this->derivingMap);
         $definition = $collection->definition('Something', 'Person');
 
         $derivings = $definition->derivings();
@@ -617,7 +604,7 @@ namespace Something;
 data DoSomething = DoSomething { string \$name, ?int \$age } deriving (Command);
 CODE;
 
-        $collection = parse($this->createDefaultFile($contents), $this->derivingsMap);
+        $collection = parse($this->createDefaultFile($contents), $this->derivingMap);
         $definition = $collection->definition('Something', 'DoSomething');
 
         $derivings = $definition->derivings();
@@ -637,7 +624,7 @@ namespace Something;
 data DoSomething = DoSomething { string \$name, ?int \$age } deriving (Command:'do-something');
 CODE;
 
-        $collection = parse($this->createDefaultFile($contents), $this->derivingsMap);
+        $collection = parse($this->createDefaultFile($contents), $this->derivingMap);
         $definition = $collection->definition('Something', 'DoSomething');
 
         $derivings = $definition->derivings();
@@ -659,7 +646,7 @@ namespace Something;
 data DoSomething = DoSomething { string \$name, ?int \$age } deriving (Command:do-something);
 CODE;
 
-        parse($this->createDefaultFile($contents), $this->derivingsMap);
+        parse($this->createDefaultFile($contents), $this->derivingMap);
     }
 
     /**
@@ -674,7 +661,7 @@ namespace Something;
 data Person = Person { string \$name, ?int \$age } deriving Equals;
 CODE;
 
-        parse($this->createDefaultFile($contents), $this->derivingsMap);
+        parse($this->createDefaultFile($contents), $this->derivingMap);
     }
 
     /**
@@ -689,7 +676,7 @@ namespace Something;
 data Person = Person { string \$name, ?int \$age } deriving (Unknown);
 CODE;
 
-        parse($this->createDefaultFile($contents), $this->derivingsMap);
+        parse($this->createDefaultFile($contents), $this->derivingMap);
     }
 
     /**
@@ -703,7 +690,7 @@ data Person = Person { string \$name, ?int \$age } where
     | strlen(\$name) === 0 => "Name too short";
 CODE;
 
-        $collection = parse($this->createDefaultFile($contents), $this->derivingsMap);
+        $collection = parse($this->createDefaultFile($contents), $this->derivingMap);
         $definition = $collection->definition('Something', 'Person');
 
         $conditions = $definition->conditions();
@@ -726,7 +713,7 @@ data Person = Person { string \$name, ?int \$age } where
     | strlen(\$name) === 0 => 'Name too short';
 CODE;
 
-        $collection = parse($this->createDefaultFile($contents), $this->derivingsMap);
+        $collection = parse($this->createDefaultFile($contents), $this->derivingMap);
         $definition = $collection->definition('Something', 'Person');
 
         $conditions = $definition->conditions();
@@ -749,7 +736,7 @@ data Person = Person { string \$name, int \$age } where
     | \$age < 18 => "Too young";
 CODE;
 
-        $collection = parse($this->createDefaultFile($contents), $this->derivingsMap);
+        $collection = parse($this->createDefaultFile($contents), $this->derivingMap);
         $definition = $collection->definition('Something', 'Person');
 
         $conditions = $definition->conditions();
@@ -779,7 +766,7 @@ data Person = Person { string \$name, ?int \$age } where
     strlen(\$name) === 0 => "Name too short";
 CODE;
 
-        parse($this->createDefaultFile($contents), $this->derivingsMap);
+        parse($this->createDefaultFile($contents), $this->derivingMap);
     }
 
     /**
@@ -795,7 +782,7 @@ data Person = Person { string \$name, ?int \$age } where
     | strlen(\$name) === 0 => Name too short;
 CODE;
 
-        parse($this->createDefaultFile($contents), $this->derivingsMap);
+        parse($this->createDefaultFile($contents), $this->derivingMap);
     }
 
     /**
@@ -811,7 +798,7 @@ data Person = Person { string \$name, ?int \$age } where
     \ strlen(\$name) === 0 => Name too short;
 CODE;
 
-        parse($this->createDefaultFile($contents), $this->derivingsMap);
+        parse($this->createDefaultFile($contents), $this->derivingMap);
     }
 
     /**
@@ -829,7 +816,7 @@ data Person = Person { string \$name, int \$age } | Chef { string \$name, } wher
         | strlen(\$name) === 0 => "Name too short";
 CODE;
 
-        $collection = parse($this->createDefaultFile($contents), $this->derivingsMap);
+        $collection = parse($this->createDefaultFile($contents), $this->derivingMap);
         $definition = $collection->definition('Something', 'Person');
 
         $conditions = $definition->conditions();
@@ -865,7 +852,7 @@ data Person = Person { string \$name, int \$age } | Chef { string \$name } where
         | strlen(\$name) === 0 => "Name too short";
 CODE;
 
-        $collection = parse($this->createDefaultFile($contents), $this->derivingsMap);
+        $collection = parse($this->createDefaultFile($contents), $this->derivingMap);
         $definition = $collection->definition('Something', 'Person');
 
         $conditions = $definition->conditions();
@@ -895,7 +882,7 @@ data Person = Person { string \$name, ?int \$age } deriving (FromArray, ToArray)
         | isset(\$age) && \$age < 22 => 'Too young';
 CODE;
 
-        $collection = parse($this->createDefaultFile($contents), $this->derivingsMap);
+        $collection = parse($this->createDefaultFile($contents), $this->derivingMap);
         $definition = $collection->definition('Something', 'Person');
 
         $derivings = $definition->derivings();
